Generate TypeScript definitions from Cesium source code
The fundamental approach to automated Typescript definitions creation is to
actually compile our code with TypeScript and have the compiler only output
the definition files instead of the compiled code. This has been a feature
of TypeScript for quite some time, but was made much more robust in version
3.7 of compiler because it can now infer additional type information from
JSDoc.

This change is the first step, which simply runs `Source/Cesium.js`
through the TS compiler and outputs the resulting definition files to
`Build/TypeScript` (which will eventually live along side other Cesium
files in Source.

There is still a lot of work to be done, but this initial change has 3
components

1. `tsconfig.declaration.json` is a TS configuration file that instructs
the compiler to process `Source/Cesium.js` and output TS definitions
for all files it uses to `Build/TypeScript`.

2. npm script, `typescript-definitions` which actually runs the compiler
as described in step 1.

3. npm script, `validate-typescript-definitions` which runs the compiler
on the generate TS definition files to ensure they are valid TypeScript.

There are errors for both steps 2 and 3 at this point, but this gives us a
base to start working on additional commits to fix it.