Minor tweaks to remove final TS compiler errors
1. `when`'s documentation was confusing the TS compiler because it has it's
own private `Promise` implementation. Since `when` is not in our JSDoc, it
was easiest to modify `when` documentation comment for now.

2. `createPropertyDescriptor` is private but TS' type inference thinks it
was trying to export a private `set` method from the module itself, which
maybe be true but simply do to lack of proper JSDoc.  Simply setting it
to return any works well enough for now.